Refactor: [BADA-68] SecurityConfig 위치 변경 #29
