+++ src/main/java/com/TwoSeaU/BaData/domain/auth/config/SecurityConfig.java
@@ -1,0 +1,46 @@
+package com.TwoSeaU.BaData.domain.auth.config;
+
+import com.TwoSeaU.BaData.domain.auth.jwt.JwtFilter;
+import com.TwoSeaU.BaData.domain.auth.jwt.handler.JwtAccessDeniedHandler;
+import com.TwoSeaU.BaData.domain.auth.jwt.handler.JwtAuthenticationEntryPointHandler;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+@EnableWebSecurity
+@RequiredArgsConstructor
+@Configuration
+public class SecurityConfig {
+
+    private final JwtAuthenticationEntryPointHandler authenticationEntryPointHandler;
+    private final JwtAccessDeniedHandler accessDeniedHandler;
+    private final JwtFilter jwtFilter;
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        http
+                .csrf(AbstractHttpConfigurer::disable)
+                .formLogin(AbstractHttpConfigurer::disable)
+                .httpBasic(AbstractHttpConfigurer::disable)
+                .sessionManagement(sessions -> sessions.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+                .exceptionHandling((exception)-> exception.authenticationEntryPoint(authenticationEntryPointHandler))
+                .exceptionHandling((exception)-> exception.accessDeniedHandler(accessDeniedHandler))
+
+                .authorizeHttpRequests((requests) ->
+                        requests
+                                .requestMatchers("/swagger", "/swagger-ui.html", "/swagger-ui/**", "/api-docs", "/api-docs/**", "/v3/api-docs/**").permitAll()
+                                .requestMatchers("/api/v1/auth/token/issue").permitAll()
+                                .anyRequest().authenticated())
+
+                .addFilterBefore(jwtFilter,
+                        UsernamePasswordAuthenticationFilter.class);
+
+        return http.build();
+    }
+
+}
